requirements.txt
asgiref==3.4.1
attrs==21.2.0
Django==3.2.8
django-environ==0.8.1
django-filter==21.1
djangorestframework==3.12.4
environ==1.0
iniconfig==1.1.1
install==1.3.4
packaging==21.0
pluggy==1.0.0
psutil==5.8.0
psycopg2-binary==2.9.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
pytest-django==4.4.0
pytz==2021.3
schedule==1.1.0
sqlparse==0.4.2
toml==0.10.2
